Cleanup connection tracking logic
