+++ main.py
@@ -2,87 +2,12 @@ import os
 
 from src.checker import Checker
 from scapy.all import sniff
-from scapy.layers.inet import IP, TCP
-from scapy.layers.inet6 import IPv6
 from src.logging.logger import Logger
-from time import time
 
 checker = Checker()
 
-class TcpConnection:
-    state: str
-    begin: float
-    initiator: str
-    port: int
-    finisher: str | None
-    src_bytes: int
-    dst_bytes: int
-
-    def __init__(self, initiator: str, port: int):
-        self.state = 'initial'
-        self.begin = time()
-        self.initiator = initiator
-        self.port = port
-        self.finisher = None
-        self.src_bytes = 0
-        self.dst_bytes = 0
-
-tcp_connections = dict[TcpConnection]()
-
 
 def packet_handler(packet):
-    if TCP in packet:
-        if IP in packet:
-            src, dst = packet[IP].src, packet[IP].dst
-        else:
-            src, dst = packet[IPv6].src, packet[IPv6].dst
-        sport, dport = packet[TCP].sport, packet[TCP].dport
-        # unique key for this TCP connection
-        key = (min(src, dst), max(src, dst), min(sport, dport), max(sport, dport))
-        if key in tcp_connections:
-            connection = tcp_connections[key]
-        else:
-            connection = TcpConnection(src, dport)
-            tcp_connections[key] = connection
-        if src == connection.initiator:
-            connection.src_bytes += len(bytes(packet))
-        else:
-            connection.dst_bytes += len(bytes(packet))
-        if connection.state == 'initial':
-            if src == connection.initiator and packet[TCP].flags == 'S':
-                connection.state = 'syn'
-            else:
-                # ignore
-                connection.state = 'ack'
-        elif connection.state == 'syn':
-            if dst == connection.initiator and packet[TCP].flags == 'SA':
-                connection.state = 'synack'
-            else:
-                connection.state = 'ack'
-        elif connection.state == 'synack':
-            if src == connection.initiator and packet[TCP].flags == 'A':
-                print(f'3-way handshake complete for {src} <-> {dst}')
-                connection.state = 'ack'
-            else:
-                connection.state = 'ack'
-        elif connection.state == 'ack':
-            if packet[TCP].flags == 'FA':
-                connection.finisher = src
-                connection.state = 'fin'
-        elif connection.state == 'fin':
-            if dst == connection.finisher and packet[TCP].flags == 'FA':
-                connection.state = 'finack'
-        elif connection.state == 'finack':
-            if src == connection.finisher and packet[TCP].flags == 'A':
-                # connection was terminated correctly
-                duration = time() - connection.begin
-                print(f'Connection terminated for {src} <-> {dst}, duration {duration}, src_bytes {connection.src_bytes}, dst_bytes {connection.dst_bytes}, flag SF')
-                del tcp_connections[key]
-        if 'R' in packet[TCP].flags:
-            duration = time() - connection.begin
-            print(f'Connection reset for {src} <-> {dst}, duration {duration}, src_bytes {connection.src_bytes}, dst_bytes {connection.dst_bytes}, flag REJ')
-            del tcp_connections[key]
-        # print(packet[TCP])
     checker.check(packet)
     Logger.debug(f"Captured Packet: {packet.summary()}\n")
